WebSockets are now working properly, the ScoreController is currently just a "reflector" to publish the incoming messages.
Changed Player's Game View to include player ID for both
the local player and their opponent.

Score changes now send websocket messages with the new score
and player ID for that score.

Changed terminology from "Passing Tests" as the score to "Stories Done"
(this is because in the game, you can bypass getting tests passing by taking on risk)