(After 4 month hiatus)
Deck (temporarily) using an "Identity" shuffler (i.e., does not shuffle) so we can focus on event-sourcing aspects of the code.

Game.start() has the beginnings of dealing with events from executing commands against the Player (and Deck) entities: Game is used to enqueue the events, but some of those events are generated by Player (and other entities), which "call back" into Game's enqueue().

Refactored events so that each entity has its own "tree" of sub-events: PlayerEvent is now the parent of PlayerDrewActionCard. This will make dispatching easier and makes it clear that those events apply to those entities.